se agrega input y alert a btn carrito

+++ src/components/CardProductosSearch.jsx
@@ -1,4 +1,4 @@
-import { useEffect } from "react";
+import { useState } from "react";
 import "./CardProductosSearch.css";
 import Swal from "sweetalert2";
 
@@ -12,31 +12,58 @@ function CardProductos({
   price,
   setCartCount,
   setFavoritesCount,
+  stock,
 }) {
+  const [quantitySelect, setQuantitySelect] = useState(1);
+
   const handleAddToCart = (product) => {
-    let cart = JSON.parse(localStorage.getItem("cartItems")) || [];
-
-    let existingProduct = cart.find((p) => p._id === product._id);
-
-    if (existingProduct) {
-      existingProduct.quantity += 1;
-    } else {
-      cart.push({ ...product, quantity: 1 });
-    }
-
-    localStorage.setItem("cartItems", JSON.stringify(cart));
     Swal.fire({
-      position: "top-center",
-      icon: "success",
-      title: "Se agregó el producto al carrito",
-      showConfirmButton: false,
-      timer: 1500,
+      title: "Ingrese la cantidad de productos",
+      input: "number",
+      inputAttributes: {
+        min: 1,
+        max: stock,
+      },
+      showCancelButton: true,
+      confirmButtonText: "Agregar al carrito",
+      showLoaderOnConfirm: true,
+      preConfirm: (quantitySelect) => {
+        return new Promise((resolve) => {
+          setTimeout(() => {
+            resolve(parseInt(quantitySelect));
+          }, 1000);
+        });
+      },
+    }).then((result) => {
+      if (result.isConfirmed) {
+        const quantitySelect = result.value;
+        let cart = JSON.parse(localStorage.getItem("cartItems")) || [];
+
+        let existingProduct = cart.find((p) => p._id === product._id);
+
+        if (existingProduct) {
+          existingProduct.quantity += quantitySelect;
+        } else {
+          cart.push({ ...product, quantity: quantitySelect });
+        }
+
+        localStorage.setItem("cartItems", JSON.stringify(cart));
+
+        Swal.fire({
+          position: "top-center",
+          icon: "success",
+          title: "Se agregó el producto al carrito",
+          showConfirmButton: false,
+          timer: 1500,
+        });
+
+        const cartCount = cart.reduce(
+          (count, item) => count + item.quantity,
+          0
+        );
+        setCartCount(cartCount);
+      }
     });
-
-    const cartCount = cart.reduce((count, item) => count + item.quantity, 0);
-
-    const updatedCartCount = cartCount;
-    setCartCount(updatedCartCount);
   };
 
   const addFavorite = (favorite) => {
@@ -88,17 +115,20 @@ function CardProductos({
         <div className="car-buttons btnCardPr d-flex justify-content-center">
           <button
             className="btn btn-primary me-2 w-50 btncart"
-            onClick={() => {
-              handleAddToCart({
-                image,
-                name,
-                description,
-                capacity,
-                category,
-                _id,
-                price,
-              });
-            }}
+            onClick={() =>
+              handleAddToCart(
+                {
+                  image,
+                  name,
+                  description,
+                  capacity,
+                  category,
+                  _id,
+                  price,
+                },
+                quantitySelect
+              )
+            }
           >
             Comprar
           </button>
