+++ src/pages/Carrito.jsx
@@ -45,13 +45,29 @@ function Carrito({ setCartCount }) {
   };
 
   const handleRemoveFromCart = (producto) => {
-    const newCartItems = cartItems.filter((item) => item._id !== producto._id);
-    if (newCartItems.length !== cartItems.length) {
-      setCartItems(newCartItems);
-      localStorage.setItem("cartItems", JSON.stringify(newCartItems));
-      const count = newCartItems.length;
-      setCartCount(count);
-    }
+    Swal.fire({
+      title: "Eliminar producto",
+      text: `¿Estás seguro que deseas eliminar ${producto.name} del carrito?`,
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Eliminar",
+      cancelButtonText: "Cancelar",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        const newCartItems = cartItems.filter(
+          (item) => item._id !== producto._id
+        );
+        if (newCartItems.length !== cartItems.length) {
+          setCartItems(newCartItems);
+          localStorage.setItem("cartItems", JSON.stringify(newCartItems));
+          const count = newCartItems.reduce(
+            (count, item) => count + item.quantity,
+            0
+          );
+          setCartCount(count);
+        }
+      }
+    });
   };
 
   const handlePayment = () => {
